Update to linux 6.15.1 and to distro cross-tools
Newer linux support, and adaptation to the ubuntu riscv cross-dev
packages.
Linux config could be skimmed much more, but this needs trial and
errors to obtain the minimal kernel, and I have no time for that.

Some parts of the image.sh script are not so bright and fails
silently and build a buggy kernel, don't hesitate to launch with
bash -x if qemu refuses to boot.